Use GCC's `alias` feature to access structures like a FlashString.
This avoids the need for indirection, but it does mean symbols have to be declared using extern "C" so increases the risk of collisions.
C++ mangled names appear like this: `_ZL17fstr_data_myContent` so we'd need the length to be able to determine the mangled name,
which isn't practical.
If this becomes an issue we can mangle the names ourselves in some way by modifying the FSTR_DATA_NAME macro.